+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Each pull request must add a one-line, user-facing entry under **Unreleased** in
 ### Fixed
 
 ### Removed
-
+ - Remove debug settings for database re-copy and parsing checkpoints (#PR_NUMBER)
 ### Security
 
 ## [4.6.0] - 2025-06-15